refac: 중복되는 style과 Components 파일로 분리 #35
CompleteQuestionCard, LiveQuestionCard,
ModerationQuestionCard 에서 중복으로 사용하는
style 과 Components 를 별도의 파일로 분리
