+++ frontend/host-app/src/components/Questions/CompleteQuestionCard.js
@@ -1,46 +1,10 @@
 import React from "react";
-import styled from "styled-components";
 import Card from "@material-ui/core/Card";
-import {CardContent, Typography} from "@material-ui/core";
+import {CardContent} from "@material-ui/core";
 import UserAvata from "./UserAvata.js";
-
-const QuestionHeader = styled.div`
-	display: flex;
-	justify-content: space-between;
-`;
-
-const QuestionMeta = styled.div`
-	display: flex;
-`;
-
-const QuestionInfo = styled.div`
-	display: flex;
-	flex-direction: column;
-	align-content: left;
-	margin-left: 1rem;
-`;
-
-const QuestionBody = styled.div``;
-
-function QuestionUserName({userName}) {
-	return (
-		<Typography
-			color={"textPrimary"}
-			variant={"subtitle1"}
-			style={{fontWeight: "bold"}}
-		>
-			{userName}
-		</Typography>
-	);
-}
-
-function QuestionDate({date}) {
-	return (
-		<Typography color={"textSecondary"} variant={"body1"}>
-			{`${date.getFullYear()}.${date.getMonth()}.${date.getDay()}`}
-		</Typography>
-	);
-}
+import {QuestionHeader, QuestionBody, QuestionInfo, QuestionMeta} from "./QuestionStyle";
+import QuestionDate from "./QuestionDate";
+import QuestionUserName from "./QuestionUserName";
 
 function CompleteQuestionCard(props) {
 	return (
@@ -54,9 +18,7 @@ function CompleteQuestionCard(props) {
 							<QuestionDate {...props} />
 						</QuestionInfo>
 					</QuestionMeta>
-
 				</QuestionHeader>
-
 				<QuestionBody>{props.question}</QuestionBody>
 			</CardContent>
 		</Card>
