+++ frontend/host-app/src/components/Questions/QuestionDate.js
@@ -1,0 +1,12 @@
+import {Typography} from "@material-ui/core";
+import React from "react";
+
+function QuestionDate({date}) {
+	return (
+		<Typography color={"textSecondary"} variant={"body1"}>
+			{`${date.getFullYear()}.${date.getMonth()}.${date.getDay()}`}
+		</Typography>
+	);
+}
+
+export default QuestionDate;
