+++ frontend/host-app/src/components/Questions/QuestionUserName.js
@@ -1,0 +1,16 @@
+import {Typography} from "@material-ui/core";
+import React from "react";
+
+function QuestionUserName({userName}) {
+	return (
+		<Typography
+			color={"textPrimary"}
+			variant={"subtitle1"}
+			style={{fontWeight: "bold"}}
+		>
+			{userName}
+		</Typography>
+	);
+}
+
+export default QuestionUserName;
